similarity/requirements.txt
chroma-hnswlib==0.7.3
chromadb==0.5.0
langchain==0.1.16
langchain-community==0.0.34
langchain-core==0.1.45
langchain-text-splitters==0.0.1
langsmith==0.1.50
python-dotenv==1.0.1
pypdf==4.2.0
huggingface-hub==0.22.2
uvicorn==0.29.0
fastapi==0.110.1
